plugin: move into package
When the definitions moved here live in a `main` package, they cannot be
imported by anyone else. Since the stuff they _use_ lives in an `internal`
package, you cannot defined them in your own code.

Hence, there's no way to combine this plugin with another one, since you
cannot create the new cmd's code using the code here.

With this change, the same thing that happens in cmd/opa-istio-plugin/main.go
could happen in another main.go that also registers another plugin.
