+++ cmd/opa-istio-plugin/main.go
@@ -8,29 +8,14 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/open-policy-agent/opa-istio-plugin/internal"
+	"github.com/open-policy-agent/opa-istio-plugin/plugin"
 	"github.com/open-policy-agent/opa/cmd"
-	"github.com/open-policy-agent/opa/plugins"
 	"github.com/open-policy-agent/opa/runtime"
 )
 
-// Factory defines the interface OPA uses to instantiate a plugin.
-type Factory struct{}
-
-// New returns the object initialized with a valid plugin configuration.
-func (Factory) New(m *plugins.Manager, config interface{}) plugins.Plugin {
-	return internal.New(m, config.(*internal.Config))
-}
-
-// Validate returns a valid configuration to instantiate the plugin.
-func (Factory) Validate(m *plugins.Manager, config []byte) (interface{}, error) {
-	return internal.Validate(m, config)
-}
-
 func main() {
-
-	runtime.RegisterPlugin("envoy.ext_authz.grpc", Factory{}) // for backwards compatibility
-	runtime.RegisterPlugin(internal.PluginName, Factory{})
+	runtime.RegisterPlugin("envoy.ext_authz.grpc", plugin.Factory{}) // for backwards compatibility
+	runtime.RegisterPlugin(plugin.PluginName, plugin.Factory{})
 
 	if err := cmd.RootCommand.Execute(); err != nil {
 		fmt.Println(err)
